# refactor: use collections (Map and Set) for storing data more efficiently