I've tried to keep each commit to just converting one thing so hopefully this will make it not too difficult to review the changes commit by commit.

Benchmarking:

| | total | loadSave | writeFileSaveAs | fogOfWar | renderObjects | addMapLayers | updateRadioactiveData | 
|:---|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
| before | 169.1s | 60.1s | 53.7s | 2.6s | 7.9s | 44.8s | <0.01s |
| after  | 163.3s | 56.9s | 51.7s | 2.6s | 7.3s | 44.7s | <0.01s |
| change | -3% | -5% | -4% | 0% | -8% | 0% | 0% |